# bx-super-app-backend/spec/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Super APP",
    "description": "Super APP Services",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
...
      "ClientUpdatedModel": {
        "type": "object",
        "properties": {
          "timestamp": { "type": "string" },
          "operation": { "type": "string" },
          "payload": { "type": "object" }
        },
        "required": ["timestamp", "operation", "payload"]
      },
...
      "createPopup": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "image": { "type": "string" },
          "show": { "type": "boolean" }
        },
        "required": ["title", "description", "image", "show"]
      },
      "updateWhatsappLink": {
        "type": "object",
        "properties": {
          "link": { "type": "string" }
        },
        "required": ["link"]
      },
...
    }
  },
  "paths": {
...
    "/api/pyme2c/backend/v1/legacy/client-integration": {
      "post": {
        "operationId": "LegacyController_ClientIntegration",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ClientUpdatedModel" }
            }
          }
        },
        "responses": { "201": { "description": "" } }
      }
    },
...
    "/api/pyme2c/backend/v1/popup": {
      "get": {
        "operationId": "PopupController_getPopup",
        "summary": "Get Popup",
        "parameters": [],
        "responses": {
          "200": { "description": "Get Popup" },
          "401": { "description": "Unauthorized" }
        },
        "tags": ["Popup"],
        "security": [{ "bearer": [] }]
      },
      "post": {
        "operationId": "PopupController_createPopup",
        "summary": "Get Popup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/createPopup" }
            }
          }
        },
        "responses": {
          "201": { "description": "Create Popup" },
          "401": { "description": "Unauthorized" }
        },
        "tags": ["Popup"],
        "security": [{ "bearer": [] }]
      }
    },
    "/api/pyme2c/backend/v1/popup/find-one": {
      "get": {
        "operationId": "PopupController_getOnePopup",
        "summary": "Get Popup",
        "parameters": [],
        "responses": {
          "200": { "description": "Get Popup" },
          "401": { "description": "Unauthorized" }
        },
        "tags": ["Popup"],
        "security": [{ "bearer": [] }]
      }
    },
    "/api/pyme2c/backend/v1/popup/{id}": {
      "put": {
        "operationId": "PopupController_updatePopup",
        "summary": "Update Popup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/createPopup" }
            }
          }
        },
        "responses": {
          "200": { "description": "Update Popup" },
          "401": { "description": "Unauthorized" }
        },
        "tags": ["Popup"],
        "security": [{ "bearer": [] }]
      }
    },
...
    "/api/pyme2c/backend/v1/settings/whatsapp-link": {
      "get": {
        "operationId": "SettingsWhatssapController_getLink",
        "summary": "Get Whatsapp Link",
        "parameters": [],
        "responses": {
          "200": { "description": "Get Whatsapp Link" },
          "401": { "description": "Unauthorized" }
        },
        "tags": ["Settings"],
        "security": [{ "bearer": [] }]
      },
      "post": {
        "operationId": "SettingsWhatssapController_updateLink",
        "summary": "Update Whatsapp link",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/updateWhatsappLink" }
            }
          }
        },
        "responses": {
          "201": { "description": "Update Whatsapp link" },
          "401": { "description": "Unauthorized" }
        },
        "tags": ["Settings"],
        "security": [{ "bearer": [] }]
      }
    }
  }
}